fix(code-block): shared shell + @layer fix for backticks + bibtex chrome
Three issues surfaced on the new build:

- Inline-code backticks were still rendering: the override sat in
  `@layer base` while @tailwindcss/typography emits the backtick rule
  in `@layer utilities`. Layer cascade beats specificity, so the
  plugin won regardless. Moved the override into `@layer utilities`
  alongside the plugin; dropping `:where()` then secures the
  specificity tiebreak (0,1,2 vs the plugin's 0,1,1).
- ResearchBibTeX still wore the old slate-700 + bg-primary-600
  stickered tab. Extracted a `CodeBlockShell` component that owns
  the new chrome (rounded surface, thin neutral border, lowercase
  monospace language chip, ghost copy button) and threaded both
  Pre.tsx and ResearchBibTeX through it.
- Code-block body padding now explicitly `py-3 px-4` so the body
  matches the header strip's rhythm; the inherited 1em padding
  from prism-themes.css was visually airy next to the tight header.
